make_watershed_bem function moved to mne.bem
